Update: Removes unwanted code, adds exceptions

+++ src/main/java/inventorymanagement/service/ItemServiceImpl.java
@@ -7,6 +7,8 @@ import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import inti.ws.spring.exception.client.BadRequestException;
+import inti.ws.spring.exception.client.NotFoundException;
 import inventorymanagement.constants.Constants;
 import inventorymanagement.dao.ItemDaoInterface;
 import inventorymanagement.entities.Item;
@@ -26,7 +28,11 @@ public class ItemServiceImpl implements ItemServiceInterface {
 
   @Override
   @Transactional
-  public ItemModel addItem(IncomingItemModel itemModel) {
+  public ItemModel addItem(IncomingItemModel itemModel) throws BadRequestException {
+    
+    if (itemModel.getProductTag() == null || itemModel.getProductTag().equals("") || itemModel.getProductId() <= 0) {
+      throw new BadRequestException("Required parameters are either missing or invalid");
+    }
 
     Product product = new Product(itemModel.getProductId());
     String tag = itemModel.getProductTag();
@@ -34,6 +40,7 @@ public class ItemServiceImpl implements ItemServiceInterface {
 
     Item item = new Item(product, tag, available);
     itemDaoImpl.save(item);
+    itemDaoImpl.refresh(item);
 
     ItemModel im = new ItemModel(item, tag, Constants.ITEM_CREATED_MESSAGE);
     return im;
@@ -41,19 +48,32 @@ public class ItemServiceImpl implements ItemServiceInterface {
 
   @Override
   @Transactional
-  public ItemModel updateItem(int id, IncomingItemModel itemModel) {
-    Item item = itemDaoImpl.getById(id);
+  public ItemModel updateItem(int id, IncomingItemModel itemModel) throws BadRequestException, NotFoundException {
+    
+    if (itemModel.getProductTag() == null || itemModel.getProductTag().equals("") || itemModel.getProductId() <= 0) {
+      throw new BadRequestException("Required parameters are either missing or invalid");
+    }
+    
+    try {
+      Item item = itemDaoImpl.getById(id);
 
-    item = itemServiceUtils.mapFromUpdateItem(itemModel, item);
-    itemDaoImpl.update(item);
-
-    ItemModel im = new ItemModel(item, Constants.ITEM_UPDATED_MESSAGE);
-    return im;
+      item = itemServiceUtils.mapFromUpdateItem(itemModel, item);
+      itemDaoImpl.update(item);
+      itemDaoImpl.flush();
+      itemDaoImpl.refresh(item);
+      ItemModel im = new ItemModel(item, Constants.ITEM_UPDATED_MESSAGE);
+      return im;
+    } catch(Exception e) {
+      throw new NotFoundException("Item with given ID does not exist");
+    }
   }
 
   @Override
   @Transactional
-  public void deleteItem(int id) {
+  public void deleteItem(int id) throws BadRequestException {
+    if (id <=0) {
+      throw new BadRequestException("Required parameters are either missing or invalid");
+    }
     Item item = new Item(id);
     itemDaoImpl.delete(item);
   }
@@ -67,14 +87,27 @@ public class ItemServiceImpl implements ItemServiceInterface {
 
   @Override
   @Transactional
-  public ItemModel getItemById(int id) {
-    Item item = itemDaoImpl.getById(id);
-    return itemServiceUtils.mapItemToModel(item);
+  public ItemModel getItemById(int id) throws BadRequestException, NotFoundException {
+    if (id <=0) {
+      throw new BadRequestException("Required parameters are either missing or invalid");
+    }
+    try {
+      Item item = itemDaoImpl.getById(id);
+      return itemServiceUtils.mapItemToModel(item);
+    }
+    catch(Exception e) {
+      throw new NotFoundException("Item with given ID does not exist");
+    }
   }
 
   @Override
   @Transactional
-  public int getCountItem(int id) {
+  public int getCountItem(int id) throws BadRequestException {
+    if (id <=0) {
+      throw new BadRequestException("Required parameters are either missing or invalid");
+    }
     return itemDaoImpl.getCountByProductId(id);
   }
+  
+  
 }
